useCreateComment used for both useForMilestone and useComments
