+++ assets/js/features/CommentSection/utils/index.tsx
@@ -1,1 +1,3 @@
 export { useEditComment } from "./useEditComment";
+export { useCreateComment } from "./useCreateComment";
+export * from "./commentParser";
